changed output of sqrt to int, removes type conversion.

+++ problem_7/python/problem7_3.py
@@ -4,7 +4,7 @@ import math
 
 def is_prime(divided):
   divisor = 3
-  sqrt_divided = math.sqrt(divided)
+  sqrt_divided = int(math.sqrt(divided))
   while divisor <= sqrt_divided:
     if divided % divisor == 0:
       return False
